backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
sqlalchemy
pydantic
apscheduler
google-genai
python-dotenv
aiosqlite
greenlet
